+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * get_client as macro
 * test get_manifest3
+* add test domain
 
 ## [0.3.2] - 2023-12-14